+++ bamt/nodes/conditional_mixture_gaussian_node.py
@@ -5,7 +5,7 @@ import numpy as np
 from gmr import GMM
 from pandas import DataFrame
 
-from bamt.utils.MathUtils import component
+from bamt.utils.math_utils import component
 from .base import BaseNode
 from .schema import CondMixtureGaussParams
