+++ tests/test_graph_analyzer.py
@@ -4,7 +4,7 @@ import unittest
 from bamt.builders.builders_base import VerticesDefiner
 from bamt.networks.discrete_bn import DiscreteBN
 from bamt.nodes.discrete_node import DiscreteNode
-from bamt.utils import GraphUtils
+from bamt.utils import graph_utils
 
 logging.getLogger("builder").setLevel(logging.CRITICAL)
